+++ __tests__/ImageHelperTest.php
@@ -3,50 +3,62 @@ use PHPUnit\Framework\TestCase;
 use \X\Util\ImageHelper;
 use \X\Util\FileHelper;
 
-const TMPDIR = __DIR__ . '/tmp';
-const INDIR = __DIR__ . '/input';
-const OUTDIR = __DIR__ . '/output';
-
 final class ImageHelperTest extends TestCase {
+  const TMPDIR = __DIR__ . '/tmp';
+  const INDIR = __DIR__ . '/input';
+  const OUTDIR = __DIR__ . '/output';
+
   public static function setUpBeforeClass(): void {
     // During testing, files in the input directory are overwritten, so reset the input directory before testing.
-    FileHelper::delete(TMPDIR);
-    FileHelper::copyDirectory(INDIR, TMPDIR);
+    FileHelper::delete(self::TMPDIR);
+    FileHelper::copyDirectory(self::INDIR, self::TMPDIR);
   }
 
   public function testWriteFirstFrameOfGifInASeparateFile(): void {
-    $src = TMPDIR . '/animated.gif';
-    $dest = OUTDIR .  '/first-frame-of-gif.gif';
+    if (!class_exists('Imagick'))
+      $this->markTestSkipped('Imagick extension is not installed.');
+    $src = self::TMPDIR . '/animated.gif';
+    $dest = self::OUTDIR .  '/first-frame-of-gif.gif';
     ImageHelper::extractFirstFrameOfGif($src, $dest);
     $this->assertSame(ImageHelper::getNumberOfGifFrames($dest), 1);
   }
 
   public function testWriteFirstFrameOfGifInSameFile(): void {
-    $src = TMPDIR . '/animated.gif';
+    if (!class_exists('Imagick'))
+      $this->markTestSkipped('Imagick extension is not installed.');
+    $src = self::TMPDIR . '/animated.gif';
     ImageHelper::extractFirstFrameOfGif($src);
     $this->assertSame(ImageHelper::getNumberOfGifFrames($src), 1);
   }
 
   public function testGetNumberOfFramesInGif(): void {
-    $src = TMPDIR . '/animated2.gif';
+    if (!class_exists('Imagick'))
+      $this->markTestSkipped('Imagick extension is not installed.');
+    $src = self::TMPDIR . '/animated2.gif';
     $this->assertSame(ImageHelper::getNumberOfGifFrames($src), 19);
   }
 
   public function testGetNumberOfFramesOfAGifWithoutAnimation(): void {
-    $src = TMPDIR . '/non-animated.gif';
+    if (!class_exists('Imagick'))
+      $this->markTestSkipped('Imagick extension is not installed.');
+    $src = self::TMPDIR . '/non-animated.gif';
     $this->assertSame(ImageHelper::getNumberOfGifFrames($src), 1);
   }
 
   public function testWriteAllPagesOfPdfAsImage(): void {
-    $src = TMPDIR . '/sample.pdf';
-    $dest = OUTDIR .  '/pdf.jpg';
+    if (!class_exists('Imagick'))
+      $this->markTestSkipped('Imagick extension is not installed.');
+    $src = self::TMPDIR . '/sample.pdf';
+    $dest = self::OUTDIR .  '/pdf.jpg';
     ImageHelper::pdf2Image($src, $dest);
     $this->assertSame(true, true);
   }
 
   public function testWriteOnlyFirstPageOfPdfAsmage(): void {
-    $src = TMPDIR . '/sample.pdf';
-    $dest = OUTDIR .  '/pdf.jpg';
+    if (!class_exists('Imagick'))
+      $this->markTestSkipped('Imagick extension is not installed.');
+    $src = self::TMPDIR . '/sample.pdf';
+    $dest = self::OUTDIR .  '/pdf.jpg';
     ImageHelper::pdf2Image($src, $dest, ['pageNumber' => 0]);
     $this->assertSame(true, true);
   }